+++ python/cudf/cudf/tests/test_dataframe.py
@@ -8128,8 +8128,13 @@ def test_sample_reproducibility(replace, random_state_lib):
     df = cudf.DataFrame({"a": cupy.arange(0, 1024)})
 
     n = 1024
-    expected = df.sample(n, replace=replace, random_state=random_state_lib(10))
-    out = df.sample(n, replace=replace, random_state=random_state_lib(10))
+    # TODO: The uint32() here is only needed for CuPy 13.2+NumPy 2 support.
+    expected = df.sample(
+        n, replace=replace, random_state=random_state_lib(np.uint32(10))
+    )
+    out = df.sample(
+        n, replace=replace, random_state=random_state_lib(np.uint32(10))
+    )
 
     assert_eq(expected, out)
 
@@ -8174,7 +8179,10 @@ def test_oversample_without_replace(n, frac, axis):
     )
 
 
-@pytest.mark.parametrize("random_state", [None, cupy.random.RandomState(42)])
+# TODO: The np.uint32() here is only temporary for CuPy 13.2 support
+@pytest.mark.parametrize(
+    "random_state", [None, cupy.random.RandomState(np.uint32(42))]
+)
 def test_sample_unsupported_arguments(random_state):
     df = cudf.DataFrame({"float": [0.05, 0.2, 0.3, 0.2, 0.25]})
     with pytest.raises(
